+++ test/functional/wallet_get_address_usage.py
@@ -31,7 +31,6 @@ from test_framework.mintlayer import block_input_data_obj
 from test_framework.wallet_cli_controller import WalletCliController
 
 import asyncio
-import sys
 
 
 class WalletGetAddressUsage(BitcoinTestFramework):