+++ test/functional/wallet_sweep_address.py
@@ -26,13 +26,12 @@ Check that:
 """
 
 from test_framework.test_framework import BitcoinTestFramework
-from test_framework.mintlayer import (ATOMS_PER_COIN, make_tx, reward_input, tx_input)
+from test_framework.mintlayer import (ATOMS_PER_COIN, make_tx, reward_input)
 from test_framework.util import assert_in, assert_equal, assert_not_in
-from test_framework.mintlayer import mintlayer_hash, block_input_data_obj
-from test_framework.wallet_cli_controller import UtxoOutpoint, WalletCliController
+from test_framework.mintlayer import block_input_data_obj
+from test_framework.wallet_cli_controller import WalletCliController
 
 import asyncio
-import sys
 import random