+++ wallet/wallet-rpc-lib/src/rpc/mod.rs
@@ -27,7 +27,9 @@ use std::{
 };
 
 use chainstate::{
-    rpc::RpcOutputValueIn, tx_verifier::check_transaction, ChainInfo, TokenIssuanceError,
+    rpc::{RpcOutputValueIn, TokenDecimals},
+    tx_verifier::check_transaction,
+    ChainInfo, TokenIssuanceError,
 };
 use common::{
     address::Address,
@@ -39,7 +41,9 @@ use common::{
         signature::inputsig::arbitrary_message::{
             produce_message_challenge, ArbitraryMessageSignature,
         },
-        tokens::{IsTokenFreezable, IsTokenUnfreezable, Metadata, TokenId, TokenTotalSupply},
+        tokens::{
+            IsTokenFreezable, IsTokenUnfreezable, Metadata, RPCTokenInfo, TokenId, TokenTotalSupply,
+        },
         Block, ChainConfig, DelegationId, Destination, GenBlock, OrderId, PoolId,
         SignedTransaction, SignedTransactionIntent, Transaction, TxOutput, UtxoOutPoint,
     },
@@ -754,6 +758,38 @@ where
                     .get_utxos(utxo_types, utxo_states, with_locked)
             })
             .await?
+    }
+
+    pub async fn get_token_infos(
+        &self,
+        token_ids: BTreeSet<TokenId>,
+    ) -> WRpcResult<BTreeMap<TokenId, RPCTokenInfo>, N> {
+        let mut result = BTreeMap::new();
+
+        // TODO: consider introducing a separate node RPC call that would fetch all token infos at once.
+        for token_id in token_ids {
+            let token_info = self
+                .node
+                .get_token_info(token_id)
+                .await
+                .map_err(RpcError::RpcError)?
+                .ok_or(RpcError::MissingTokenInfo(token_id))?;
+            result.insert(token_id, token_info);
+        }
+
+        Ok(result)
+    }
+
+    pub async fn get_tokens_decimals(
+        &self,
+        token_ids: BTreeSet<TokenId>,
+    ) -> WRpcResult<BTreeMap<TokenId, TokenDecimals>, N> {
+        Ok(self
+            .get_token_infos(token_ids)
+            .await?
+            .iter()
+            .map(|(id, info)| (*id, TokenDecimals(info.token_number_of_decimals())))
+            .collect())
     }
 
     pub async fn get_transaction(